修复：由于QQ长度过长导致值溢出

+++ src/Plugin/PCR/orm/damage.py
@@ -9,7 +9,7 @@ class Damage(BaseModel):
     stage = IntegerField()
     step = IntegerField()
     group = IntegerField()
-    member = IntegerField()
+    member = BigIntegerField()
     damage = IntegerField()
     kill = BooleanField()
